Refactor config files
