+++ src/mxcp/config/execution_engine.py
@@ -42,78 +42,16 @@ Example usage:
 """
 
 import logging
-from collections.abc import Generator
-from contextlib import contextmanager
 from pathlib import Path
-from typing import Any, cast
 
 from mxcp.config._types import SiteConfig, UserConfig
-from mxcp.sdk.executor import (
-    ExecutionContext,
-    ExecutionEngine,
-    reset_execution_context,
-    set_execution_context,
-)
+from mxcp.config.duckdb_config import create_duckdb_session_config
+from mxcp.config.execution_context import execution_context_for_init_hooks
+from mxcp.config.site_config import find_repo_root
+from mxcp.sdk.executor import ExecutionEngine
 from mxcp.sdk.executor.plugins import DuckDBExecutor, PythonExecutor
-from mxcp.sdk.executor.plugins.duckdb_plugin._types import (
-    DatabaseConfig,
-    ExtensionDefinition,
-    PluginConfig,
-    PluginDefinition,
-    SecretDefinition,
-)
-from mxcp.sdk.executor.plugins.duckdb_plugin.session import DuckDBSession
 
 logger = logging.getLogger(__name__)
-
-
-@contextmanager
-def execution_context_for_init_hooks(
-    user_config: UserConfig | None = None,
-    site_config: SiteConfig | None = None,
-    duckdb_session: DuckDBSession | None = None,
-    plugins: dict[str, Any] | None = None,
-) -> Generator[ExecutionContext | None, None, None]:
-    """
-    Context manager for setting up ExecutionContext for init hooks.
-
-    This helper function creates an ExecutionContext with the provided
-    runtime data, sets it as the current context, and automatically
-    cleans it up when done.
-
-    Args:
-        user_config: UserConfig object containing user configuration for runtime context
-        site_config: SiteConfig object containing site configuration for runtime context
-        duckdb_session: DuckDB session for runtime context
-        plugins: Plugins dict for runtime context
-
-    Yields:
-        The ExecutionContext that was created and set
-
-    Example:
-        >>> with execution_context_for_init_hooks(user_config, site_config, session) as context:
-        ...     run_init_hooks()  # init hooks have access to context
-    """
-    context = None
-    token = None
-
-    try:
-        if user_config and site_config and duckdb_session:
-            context = ExecutionContext()
-            context.set("user_config", user_config)
-            context.set("site_config", site_config)
-            context.set("duckdb_session", duckdb_session)
-            if plugins:
-                context.set("plugins", plugins)
-            token = set_execution_context(context)
-            logger.info("Set up ExecutionContext for init hooks")
-
-        yield context
-
-    finally:
-        if token:
-            reset_execution_context(token)
-            logger.info("Cleaned up ExecutionContext after init hooks")
 
 
 def create_execution_engine(
@@ -167,77 +105,20 @@ def create_execution_engine(
 
         # Get the profile name to use
         profile_name = profile or site_config["profile"]
-        project_name = site_config["project"]
+        
+        # Handle readonly override
+        db_readonly_from_config = False
+        if "profiles" in site_config:
+            site_profiles = site_config.get("profiles", {})
+            site_profile_config = site_profiles.get(profile_name, {})
+            duckdb_config = site_profile_config.get("duckdb", {})
+            db_readonly_from_config = duckdb_config.get("readonly", False) if duckdb_config else False
+        db_readonly = readonly if readonly is not None else db_readonly_from_config
 
-        # Get database configuration from site config profiles section
-        site_profiles = site_config.get("profiles", {})
-        site_profile_config = site_profiles.get(profile_name, {})
-        duckdb_config = site_profile_config.get("duckdb", {})
-
-        # Get database path and readonly setting from site config
-        db_path = duckdb_config.get("path") if duckdb_config else None
-        if not db_path:
-            db_path = f"data/db-{profile_name}.duckdb"
-        db_readonly_from_config = duckdb_config.get("readonly") if duckdb_config else False
-        db_readonly = readonly if readonly is not None else (db_readonly_from_config or False)
-
-        # Get extensions from site config (root level)
-        extensions_config = site_config.get("extensions") or []
-        extensions = []
-        for ext in extensions_config:
-            if isinstance(ext, str):
-                # Simple string extension name
-                extensions.append(ExtensionDefinition(name=ext))
-            elif isinstance(ext, dict):
-                # Extension with repo specification
-                ext_name = ext.get("name")
-                if ext_name:
-                    extensions.append(ExtensionDefinition(name=ext_name, repo=ext.get("repo")))
-
-        database_config = DatabaseConfig(
-            path=db_path, readonly=bool(db_readonly), extensions=extensions
+        # Create SDK session configuration using the shared function
+        database_config, plugins_list, plugin_config, secrets_list = create_duckdb_session_config(
+            site_config, user_config, profile_name, readonly=db_readonly
         )
-
-        # Get plugins from site config plugin array
-        plugins_list = []
-        site_plugins = site_config.get("plugin") or []
-        for plugin_def in site_plugins:
-            plugin_name = plugin_def.get("name")
-            plugin_module = plugin_def.get("module")
-            if plugin_name and plugin_module:
-                plugins_list.append(
-                    PluginDefinition(
-                        name=plugin_name,
-                        module=plugin_module,
-                        config=plugin_def.get("config"),  # References config key in user config
-                    )
-                )
-
-        # Get plugin configuration from user config
-        user_projects = user_config.get("projects") or {}
-        user_project = cast(dict[str, Any], user_projects.get(project_name) or {})
-        user_profiles = user_project.get("profiles") or {}
-        user_profile = user_profiles.get(profile_name) or {}
-        user_plugin_section = user_profile.get("plugin") or {}
-        user_plugin_configs = user_plugin_section.get("config") or {}
-
-        # Get plugins path from site config
-        site_paths = site_config.get("paths") or {}
-        plugins_path = site_paths.get("plugins") or "plugins"
-
-        plugin_config = PluginConfig(plugins_path=plugins_path, config=user_plugin_configs)
-
-        # Get secrets from user config profile
-        secrets_list = []
-        user_secrets = user_profile.get("secrets") or []
-        for secret in user_secrets:
-            secret_name = secret.get("name")
-            secret_type = secret.get("type")
-            secret_params = secret.get("parameters")
-            if secret_name and secret_type and secret_params:
-                secrets_list.append(
-                    SecretDefinition(name=secret_name, type=secret_type, parameters=secret_params)
-                )
 
         # Create and register DuckDB executor
         duckdb_executor = DuckDBExecutor(
@@ -273,24 +154,3 @@ def create_execution_engine(
         raise RuntimeError(f"Failed to create ExecutionEngine: {e}") from e
 
 
-def find_repo_root() -> Path:
-    """Find the repository root directory.
-
-    This function attempts to find the root of the current repository by looking
-    for common repository markers (.git, pyproject.toml, etc.).
-
-    Returns:
-        Path to the repository root, or current working directory if not found
-    """
-    current = Path.cwd()
-
-    # Look for common repository markers
-    markers = [".git", "pyproject.toml", "setup.py", "Cargo.toml", "package.json", "mxcp-site.yml"]
-
-    # Walk up the directory tree
-    for parent in [current] + list(current.parents):
-        if any((parent / marker).exists() for marker in markers):
-            return parent
-
-    # Fall back to current directory
-    return current
