+++ src/mxcp/evals/runner.py
@@ -3,7 +3,8 @@ import time
 from typing import Any, cast
 
 from mxcp.config._types import SiteConfig, UserConfig
-from mxcp.config.execution_engine import create_execution_engine, find_repo_root
+from mxcp.config.execution_engine import create_execution_engine
+from mxcp.config.site_config import find_repo_root
 from mxcp.endpoints._types import EndpointDefinition
 from mxcp.endpoints.loader import EndpointLoader
 from mxcp.evals.loader import discover_eval_files, load_eval_suite
