+++ tests/test_endpoint_execution.py
@@ -49,9 +49,9 @@ def site_config(test_repo_path):
 
 
 @pytest.fixture
-def execution_engine(user_config, site_config):
+def execution_engine(user_config, site_config, test_repo_path):
     """Create execution engine for tests."""
-    return create_execution_engine(user_config, site_config)
+    return create_execution_engine(user_config, site_config, repo_root=test_repo_path)
 
 
 def test_endpoint_loading(site_config, test_repo_path):
